Rewrote cs_access to be more C++ish, changed NumberList to be more C++ish and fixed some compiler warnings on 64bit systems

+++ src/protocol/bahamut.cpp
@@ -179,7 +179,7 @@ class BahamutIRCdProto : public IRCDProto
 	/* SGLINE */
 	void SendSGLine(XLine *x)
 	{
-		send_cmd(NULL, "SGLINE %d :%s:%s", x->Mask.length(), x->Mask.c_str(), x->Reason.c_str());
+		send_cmd(NULL, "SGLINE %d :%s:%s", static_cast<int>(x->Mask.length()), x->Mask.c_str(), x->Reason.c_str());
 	}
 
 	/* RAKILL */
